New: Default severity for `no-protocol-relative-urls`
Fix #3242
