+++ packages/hint-no-protocol-relative-urls/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@hint/utils-debug": "^1.0.0",
     "@hint/utils-i18n": "^1.0.0",
+    "@hint/utils-network": "^1.0.0",
     "@hint/utils-string": "^1.0.0",
     "@hint/utils-types": "^1.0.0"
   },
